Move methods to helper for reusability
